Split POSIX into clean POSIX and extended POSIX
